+++ src/lib/components/forms/NumberInput.tsx
@@ -58,15 +58,11 @@ export const NumberInput = ({
   // Design system size: md = 40px, lg = 56px
   return (
     <FormControl
-      className={`${size}-form`}
       isInvalid={!!error || status?.state === "error"}
       size={size}
       {...componentProps}
     >
-      <FormLabel
-        className={value.length ? "floating" : ""}
-        backgroundColor={labelBgColor}
-      >
+      <FormLabel className={`${size}-label`} backgroundColor={labelBgColor}>
         {label}
       </FormLabel>