+++ src/pages/[network]/codes.tsx
@@ -1,0 +1,3 @@
+import Codes from "lib/pages/codes";
+
+export default Codes;