+++ src/pages/[network]/migrate.tsx
@@ -1,0 +1,3 @@
+import Migrate from "lib/pages/migrate";
+
+export default Migrate;